pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smartbuildsim"
version = "0.1.0"
description = "Synthetic smart building simulation toolkit with CLI workflows"
authors = [{ name = "SmartBuildSim Team" }]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "joblib>=1.3",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "pydantic>=1.10,<3",
    "PyYAML>=6.0",
    "scikit-learn>=1.2",
    "typer[all]>=0.9",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


[project.urls]
homepage = "https://example.com/smartbuildsim"
documentation = "https://example.com/smartbuildsim/docs"

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]

[project.scripts]
smartbuildsim = "smartbuildsim.cli.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["E203", "E501", "B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers --strict-config --cov=smartbuildsim --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["smartbuildsim"]
